requirements.txt.txt
streamlit
pandas
numpy
rdkit
scikit-learn
altair
Pillow